refactor

+++ blog.py
@@ -70,11 +70,9 @@ def logout():
 @app.route('/uploads/', methods=['GET', 'POST'])
 def uploads():
     user = session.get('user')
-    if not user:
-        return redirect(url_for('login'))
-    if user['login'] != os.getenv('allowed_user'):
-        return redirect(url_for('login')), session.pop('user', None)
-    if request.method == 'POST':
+    if user and user['login'] == os.getenv('allowed_user') and request.method == 'GET':
+        return render_template('uploads.html')
+    if user and user['login'] == os.getenv('allowed_user') and request.method == 'POST':
         if 'file' not in request.files:
             return "No file part", 400
         file = request.files['file']
@@ -86,23 +84,17 @@ def uploads():
             filename = file.filename
             file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
             return f"File uploaded successfully. View: <a href='/posts/{filename.rsplit('.', 1)[0]}/'>{filename}</a>", 200
-    return render_template('uploads.html')
 
 @app.route('/delete/<name>', methods=['GET'])
 def delete(name):
     user = session.get('user')
-    if not user:
-        return redirect(url_for('login'))
-    if user['login'] != os.getenv('allowed_user'):
-        return redirect(url_for('login')), session.pop('user', None)
-    if request.method == 'GET':
+    if user and user['login'] == os.getenv('allowed_user') and request.method == 'GET':
         file_path = os.path.join(app.config['UPLOAD_FOLDER'], name)
         if os.path.exists(file_path):
             os.remove(file_path)
             return redirect(url_for('index'))
         else:
             return "File not found", 404
-    
 
 @app.route('/posts/')
 def posts():
